+++ CHANGES.rst
@@ -1,10 +1,20 @@
 Changelog
 =========
 
-3.3.1 (unreleased)
-------------------
+4.0.0a1 (unreleased)
+--------------------
 
-- Nothing changed yet.
+- Register our own ``IImageScaleFactory`` factory for persistent tiles.
+  This fixes scaling images on tiles when the tile does not have the workaround of defining a property for the image field.
+  [maurits]
+
+- Register our ``AnnotationStorage`` as ``IImageScaleStorage`` factory.
+  We require ``plone.namedfile >= 6.0.0a2`` for this.
+  Fixes `issue 49 <https://github.com/plone/plone.app.tiles/issues/48>.  [maurits]
+
+- Drop support for Python 2 and Plone 5.1.
+  Plone 5.2 and 6.0 are supported, on Python 3.
+  See `issue 49 <https://github.com/plone/plone.app.tiles/issues/49>.  [maurits]
 
 
 3.3.0 (2022-02-04)